Put a warning in the README that this is a fork that's diverged from the original jekyll-feed.
I'm going to start making more drastic customizations to this fork, to experiment with things I need for the RSS feed of www.kronopath.com. If the original pull requests I made to jekyll-feed get picked up, that's nice, and maybe I'll go back to using the mainline branch. If not, that's also okay.